# fix: remove needless pointers for increase performance