--- a/contraction.go
+++ b/contraction.go
@@ -101,7 +101,8 @@ func (graph *Graph) processIncidentEdges(vertex *Vertex, pmax float64) {
 	if len(outcomingEdges) == 0 {
 		return
 	}
-	batchShortcuts := make([]*ShortcutPath, 0, len(incomingEdges)*len(outcomingEdges))
+
+	batchShortcuts := make([]ShortcutPath, 0)
 
 	previousOrderPos := int64(vertex.orderPos - 1)
 	for _, u := range incomingEdges {
@@ -128,17 +129,18 @@ func (graph *Graph) processIncidentEdges(vertex *Vertex, pmax float64) {
 				outVertexPtr.distance.previousSourceID != inVertex { // Optional condition: if previous shortestPathsWithMaxCost(...) call has changed shortest path tree
 
 				// Collect needed shortcuts
-				batchShortcuts = append(batchShortcuts, &ShortcutPath{From: inVertex, To: outVertex, Via: vertex.vertexNum, Cost: neighborsWeights})
+				batchShortcuts = append(batchShortcuts, ShortcutPath{From: inVertex, To: outVertex, Via: vertex.vertexNum, Cost: neighborsWeights})
 			}
 		}
 	}
+
 	graph.insertShortcuts(batchShortcuts)
 }
 
 // insertShortcuts Creates (or updates: it depends on conditions) multiple shortcuts in graph structure
-func (graph *Graph) insertShortcuts(shortcuts []*ShortcutPath) {
-	for i := range shortcuts {
-		d := shortcuts[i]
+func (graph *Graph) insertShortcuts(batchShortcuts []ShortcutPath) {
+	for i := range batchShortcuts {
+		d := batchShortcuts[i]
 		graph.createOrUpdateShortcut(d.From, d.To, d.Via, d.Cost)
 	}
 }